select(): Add a test that invokes the function and prints the result
It turns out the previous code failed with error:

    thread 'select::tests::select_extracts_a_database_record_as_yaml'
    panicked at 'called `Result::unwrap()` on an `Err` value:
    SqliteFailure(Error { code: Unknown, extended_code: 1 }, Some("near
    \":table\": syntax error"))', src/select.rs:11:9

Seems that I can't use named parameters to pass the table name to
`FROM`. Use string formatting for that instead.

Converted the named parameter array to a `rusqlite::named_params!` macro
when I read about it, as that seems a bit more concise.