select(): Remove parameterized primary key column name
It turns out that adding the primary key column name to the query via a
prepared parameter was causing it to be set as a literal instead of a
column name. The resulting SQL was:

    SELECT
      *
    FROM test
    WHERE 'id' = '1';

instead of:

    SELECT
      *
    FROM test
    WHERE "id" = '1';

When the `id` column is hard-coded, this works correctly, producing the
right SQL and extracting the correct record.

We'll need to update this to add the column name via the format string
like we've done with the table name so we can make it dynamic.