select(): Build a `yaml_rust::yaml::Hash` for each row
Instead of my earlier `std::collections::HashMap`, build the type we
want to return directly in the row closure.

Required copying some data like the column names, but I don't see a way
around that. I suppose we could make `yaml_rust::Yaml`s from the column
headers in the row closure too instead of copying an existing column
header that we otherwise never use.

This nearly passes the test I have for this function, but it includes
the `id` column in the `Yaml` hash. The primary key column will need to
be removed.