select: Fix test by adding new argument
We added a new argument to the `select()` functions in
f30a4dafa5736c58bda54569ec9f46e6fdb2e200 to allow columns to be excluded
from the YAML output. Update the test to include the new argument in its
call.

I was getting this error about the `C` type parameter on the function:

    error[E0282]: type annotations needed
       --> src/select.rs:152:23
        |
    152 |             let got = select(&conn, "test", "1", None).unwrap();
        |                       ^^^^^^ cannot infer type for type parameter `C` declared on the function `select`

    For more information about this error, try `rustc --explain E0282`.

The error makes sense in a way, but it also doesn't at all, because I'm
not passing a `Some` value into the function. If it's a `None`, then
what need is there to specify a type that's never used?

I wonder if there's a way to get this to work in a way that doesn't
require me to specify the type parameter when passing `None` for
`exclude_columns`.