extract(): Rename function to `db_insert()`
The name `extract()` didn't make sense to me any more since the function
is doing more than its originally intended behaviour which was to
extract the YAML into a hash-like format for easy insertion into the
database.

I decided to give up on that behaviour since this works for what we want
it to do. Maybe this function is doing more than it should, but I'm okay
with that for the current functional requirements.